Add A-MoreOOP features
